pyproject.toml
[tool.poetry]
name = "contractmind-backend"
version = "1.0.0"
description = "ContractMind AI-powered blockchain infrastructure backend"
authors = ["ContractMind Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
web3 = "^6.11.3"
eth-account = "^0.10.0"
eth-abi = "^4.2.1"
hexbytes = "^0.3.1"
# LLM Providers
google-generativeai = "^0.3.0"
anthropic = "^0.7.7"
openai = "^1.3.7"
fastmcp = "^0.2.0"
# Database - using psycopg2 directly instead of SQLAlchemy
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
aioredis = "^2.0.1"
python-socketio = "^5.10.0"
websockets = "^12.0"
httpx = "^0.27.0"
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
pydantic-extra-types = "^2.1.0"
greenlet = "^3.2.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.12.0"
isort = "^5.13.2"
mypy = "^1.7.1"
ruff = "^0.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
